feat(list): groupListItems added
