+++ src/components/List/List.style.ts
@@ -1,5 +1,6 @@
 import { css, SerializedStyles } from '@emotion/react';
 import { Theme } from 'theme';
+import { rem } from 'theme/utils';
 
 export const wrapperStyle = ({
   width,
@@ -11,4 +12,23 @@ export const wrapperStyle = ({
   border: ${isSearchable ? 'initial' : `1px solid ${theme.utils.getColor('lightGrey', 100)}`};
   border-radius: ${isSearchable ? 'initial' : theme.spacing.xsm};
   width: ${`${width}px` || '100%'};
+`;
+
+export const listStyle = ({
+  width,
+  height,
+  isSearchable,
+}: {
+  width?: number;
+  height?: number;
+  isSearchable?: boolean;
+}) => (theme: Theme): SerializedStyles => css`
+  padding-left: 0;
+  margin-top: 0;
+  margin-bottom: 0;
+  border-radius: ${isSearchable ? 'initial' : theme.spacing.xsm};
+  width: ${width ? rem(width) : '100%'};
+  height: ${height ? rem(height) : '100%'};
+  overflow: auto;
+  overflow-x: hidden;
 `;
