+++ src/components/List/ListItem/ListItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Highlighter from 'react-highlight-words';
 import { TestProps } from 'utils/types';
 
 import { ListItemType, ListRowSize, SelectHandlerType } from '../types';
+import { renderContent } from '../utils';
 import { listItemStyle, contentStyle } from './ListItem.style';
 
 type Props = {
@@ -47,50 +47,6 @@ const ListItem = React.forwardRef<HTMLDivElement, Props>(
         handleOptionClick(content as never);
       }
     };
-    const renderContent = (content: ListItemType) => {
-      /**
-       * Check if list item is not react element because it can be
-       * and also checks if its object with property 'value'
-       * @TODO Typescript 4.4 will solve this in one constant
-       * **/
-      if (
-        searchTerm &&
-        content &&
-        !React.isValidElement(content) &&
-        typeof content === 'object' &&
-        !Array.isArray(content) &&
-        'label' in content &&
-        content?.label
-      ) {
-        return (
-          <Highlighter
-            highlightClassName="search-text"
-            highlightTag={'strong'}
-            searchWords={[searchTerm]}
-            autoEscape={true}
-            textToHighlight={content.label}
-          />
-        );
-      }
-
-      /**
-       * Check if list item is not react element because it can be
-       * and also checks if its object with property 'value'
-       * @TODO Typescript 4.4 will solve this in one constant
-       * **/
-      if (
-        content &&
-        !React.isValidElement(content) &&
-        typeof content === 'object' &&
-        !Array.isArray(content) &&
-        'label' in content &&
-        content?.label
-      ) {
-        return content.label;
-      }
-
-      return content;
-    };
 
     return (
       <div
@@ -104,7 +60,7 @@ const ListItem = React.forwardRef<HTMLDivElement, Props>(
       >
         <div css={contentStyle()}>
           {/** @TODO latest version typescript 4.4 is solving this as a constant */
-          renderContent(content)}
+          renderContent(content, searchTerm)}
         </div>
       </div>
     );
